+++ app/(public)/privacy-policy/page.tsx
@@ -1,7 +1,7 @@
-export const metadata = {
-  title: "Privacy Policy | Smriti AI",
-  description:
-    "Read the Privacy Policy of Smriti AI to learn how we collect, use, and protect your personal information. Your data privacy and security are our top priority.",
+import { generateMetadataUtil } from "@/utils/generateMetadata";
+export const metadata = generateMetadataUtil({
+  title: "Privacy Policy",
+  description: "Read the Privacy Policy of Smriti AI to learn how we collect, use, and protect your personal information. Your data privacy and security are our top priority.",
   keywords: [
     "Smriti AI Privacy Policy",
     "AI data protection",
@@ -10,8 +10,12 @@ export const metadata = {
     "data collection policy",
     "GDPR compliance",
     "AI privacy guidelines",
+    "data protection",
+    "privacy rights",
+    "secure AI platform"
   ],
-};
+  url: "https://www.smriti.live/privacy-policy",
+});
 
 export default function PrivacyPolicy() {
   return (
